# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ton SWE Challenge — API",
    "description": "Autenticação por JWT e catálogo paginado. Erros seguem a RFC 9457 e sempre trazem `code` e `requestId`.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "summary": "Autentica e abre uma sessão",
        "tags": [
          "Autenticação"
        ],
        "description": "Devolve um access token de curta duração e um refresh token de uso único. Credencial inválida, conta inexistente e conta temporariamente bloqueada produzem a mesma resposta, de propósito.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "E-mail cadastrado."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128,
                    "description": "Senha, de 8 a 128 caracteres."
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "description": "Credenciais de acesso."
              }
            }
          },
          "description": "Credenciais de acesso."
        },
        "responses": {
          "200": {
            "description": "Sessão recém-criada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "JWT de acesso, usado no cabeçalho Authorization."
                    },
                    "refreshToken": {
                      "type": "string",
                      "description": "Credencial de renovação, de uso único. Guarde com o mesmo cuidado da senha."
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991,
                      "description": "Validade do access token, em segundos."
                    },
                    "userId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "tokenType",
                    "expiresIn",
                    "userId"
                  ],
                  "additionalProperties": false,
                  "description": "Sessão recém-criada."
                }
              }
            }
          },
          "400": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "401": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "429": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "500": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/logout": {
      "post": {
        "summary": "Encerra a sessão",
        "tags": [
          "Autenticação"
        ],
        "description": "Revoga a família do refresh token apresentado. O access token em circulação permanece válido até expirar, no máximo o tempo de `expiresIn`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 512
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "description": "Credencial de renovação obtida no login ou na renovação anterior."
              }
            }
          },
          "description": "Credencial de renovação obtida no login ou na renovação anterior."
        },
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "401": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "429": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          },
          "500": {
            "description": "Erro no formato RFC 9457.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_ERROR",
                        "UNAUTHENTICATED",
                        "TOKEN_INVALID",
                        "TOKEN_EXPIRED",
                        "NOT_FOUND",
                        "INTERNAL_ERROR",
                        "INVALID_CURSOR",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "REFRESH_TOKEN_INVALID",
                        "REFRESH_TOKEN_REUSE_DETECTED",
                        "CONCURRENT_MODIFICATION",
                        "RATE_LIMIT_EXCEEDED"
                      ]
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Presente apenas em falhas de validação.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}